Updated Listing 16.01 so parser output conforms to previous pdf version
